+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # CHANGELOG
 
 ## 0.7.6dev
-* [Fix] Custom Connection using sqlalchemy now needs to be PEP 249 Compilant 
+* [Fix] Fix error when checking if custom connection was PEP 249 Compliant 
 ## 0.7.5 (2023-05-24)
 
 * [Feature] Using native DuckDB `.df()` method when using `autopandas`